requirements.txt
aniso8601==9.0.1
anyio==3.5.0
ariadne==0.14.1
azure-core==1.23.0
azure-cosmos==4.2.0
azure-functions==1.10.1
azure-functions-worker==1.1.9
certifi==2021.10.8
charset-normalizer==2.0.12
databricks-connect==7.3.*
databricks-sql-connector==1.0.1
future==0.18.2
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
grpcio==1.33.2
grpcio-tools==1.33.2
idna==3.3
protobuf==3.19.4
py4j==0.10.9
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
thrift==0.15.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
